+ Mac OS X port

+++ src/Base/PyCXX/cxx_extensions.cxx
@@ -150,15 +150,15 @@ extern "C"
 	static PyObject* call_handler (PyObject*, PyObject*, PyObject*);
 
 	// Sequence methods
-	static int sequence_length_handler(PyObject*);
+	static Py_ssize_t sequence_length_handler(PyObject*);
 	static PyObject* sequence_concat_handler(PyObject*,PyObject*);
-	static PyObject* sequence_repeat_handler(PyObject*, int);
-	static PyObject* sequence_item_handler(PyObject*, int);
-	static PyObject* sequence_slice_handler(PyObject*, int, int);
-	static int sequence_ass_item_handler(PyObject*, int, PyObject*);
-	static int sequence_ass_slice_handler(PyObject*, int, int, PyObject*);
+	static PyObject* sequence_repeat_handler(PyObject*, Py_ssize_t);
+	static PyObject* sequence_item_handler(PyObject*, Py_ssize_t);
+	static PyObject* sequence_slice_handler(PyObject*, Py_ssize_t, Py_ssize_t);
+	static int sequence_ass_item_handler(PyObject*, Py_ssize_t, PyObject*);
+	static int sequence_ass_slice_handler(PyObject*, Py_ssize_t, Py_ssize_t, PyObject*);
 	// Mapping
-	static int mapping_length_handler(PyObject*);
+	static Py_ssize_t mapping_length_handler(PyObject*);
 	static PyObject* mapping_subscript_handler(PyObject*, PyObject*);
 	static int mapping_ass_subscript_handler(PyObject*, PyObject*, PyObject*);
 
@@ -187,9 +187,9 @@ extern "C"
 	static PyObject* number_power_handler(PyObject*, PyObject*, PyObject*);
 
 	// Buffer
-	static int buffer_getreadbuffer_handler (PyObject*, int, void**);
-	static int buffer_getwritebuffer_handler (PyObject*, int, void**);
-	static int buffer_getsegcount_handler (PyObject*, int*);
+	static Py_ssize_t buffer_getreadbuffer_handler (PyObject*, Py_ssize_t, void**);
+	static Py_ssize_t buffer_getwritebuffer_handler (PyObject*, Py_ssize_t, void**);
+	static Py_ssize_t buffer_getsegcount_handler (PyObject*, Py_ssize_t*);
 	};
 
 
@@ -557,7 +557,7 @@ extern "C" PyObject* call_handler( PyObject *self, PyObject *args, PyObject *kw 
 
 
 // Sequence methods
-extern "C" int sequence_length_handler( PyObject *self )
+extern "C" Py_ssize_t sequence_length_handler( PyObject *self )
 	{
 	try
 		{
@@ -583,7 +583,7 @@ extern "C" PyObject* sequence_concat_handler( PyObject *self, PyObject *other )
 		}
 	}
 
-extern "C" PyObject* sequence_repeat_handler( PyObject *self, int count )
+extern "C" PyObject* sequence_repeat_handler( PyObject *self, Py_ssize_t count )
 	{
 	try
 		{
@@ -596,7 +596,7 @@ extern "C" PyObject* sequence_repeat_handler( PyObject *self, int count )
 		}
 	}
 
-extern "C" PyObject* sequence_item_handler( PyObject *self, int index )
+extern "C" PyObject* sequence_item_handler( PyObject *self, Py_ssize_t index )
 	{
 	try
 		{
@@ -609,7 +609,7 @@ extern "C" PyObject* sequence_item_handler( PyObject *self, int index )
 		}
 	}
 
-extern "C" PyObject* sequence_slice_handler( PyObject *self, int first, int last )
+extern "C" PyObject* sequence_slice_handler( PyObject *self, Py_ssize_t first, Py_ssize_t last )
 	{
 	try
 		{
@@ -622,7 +622,7 @@ extern "C" PyObject* sequence_slice_handler( PyObject *self, int first, int last
 		}
 	}
 
-extern "C" int sequence_ass_item_handler( PyObject *self, int index, PyObject *value )
+extern "C" int sequence_ass_item_handler( PyObject *self, Py_ssize_t index, PyObject *value )
 	{
 	try
 		{
@@ -635,7 +635,7 @@ extern "C" int sequence_ass_item_handler( PyObject *self, int index, PyObject *v
 		}
 	}
 
-extern "C" int sequence_ass_slice_handler( PyObject *self, int first, int last, PyObject *value )
+extern "C" int sequence_ass_slice_handler( PyObject *self, Py_ssize_t first, Py_ssize_t last, PyObject *value )
 	{
 	try
 		{
@@ -649,7 +649,7 @@ extern "C" int sequence_ass_slice_handler( PyObject *self, int first, int last, 
 	}
 
 // Mapping
-extern "C" int mapping_length_handler( PyObject *self )
+extern "C" Py_ssize_t mapping_length_handler( PyObject *self )
 	{
 	try
 		{
@@ -976,7 +976,7 @@ extern "C" PyObject* number_power_handler( PyObject *self, PyObject *x1, PyObjec
 	}
 
 // Buffer
-extern "C" int buffer_getreadbuffer_handler( PyObject *self, int index, void **pp )
+extern "C" Py_ssize_t buffer_getreadbuffer_handler( PyObject *self, Py_ssize_t index, void **pp )
 	{
 	try
 		{
@@ -989,7 +989,7 @@ extern "C" int buffer_getreadbuffer_handler( PyObject *self, int index, void **p
 		}
 	}
 
-extern "C" int buffer_getwritebuffer_handler( PyObject *self, int index, void **pp )
+extern "C" Py_ssize_t buffer_getwritebuffer_handler( PyObject *self, Py_ssize_t index, void **pp )
 	{
 	try
 		{
@@ -1002,7 +1002,7 @@ extern "C" int buffer_getwritebuffer_handler( PyObject *self, int index, void **
 		}
 	}
 
-extern "C" int buffer_getsegcount_handler( PyObject *self, int *count )
+extern "C" Py_ssize_t buffer_getsegcount_handler( PyObject *self, Py_ssize_t *count )
 	{
 	try
 		{
@@ -1062,30 +1062,30 @@ Py::Object PythonExtensionBase::call( const Py::Object &, const Py::Object & )
 
 
 // Sequence methods
-int PythonExtensionBase::sequence_length()
+Py_ssize_t PythonExtensionBase::sequence_length()
 	{ missing_method( sequence_length ); return -1; }
 
 Py::Object PythonExtensionBase::sequence_concat( const Py::Object & )
 	{ missing_method( sequence_concat ); return Py::Nothing(); }
 
-Py::Object PythonExtensionBase::sequence_repeat( int )
+Py::Object PythonExtensionBase::sequence_repeat( Py_ssize_t )
 	{ missing_method( sequence_repeat ); return Py::Nothing(); }
 
-Py::Object PythonExtensionBase::sequence_item( int )
+Py::Object PythonExtensionBase::sequence_item( Py_ssize_t )
 	{ missing_method( sequence_item ); return Py::Nothing(); }
 
-Py::Object PythonExtensionBase::sequence_slice( int, int )
+Py::Object PythonExtensionBase::sequence_slice( Py_ssize_t, Py_ssize_t )
 	{ missing_method( sequence_slice ); return Py::Nothing(); }
 
-int PythonExtensionBase::sequence_ass_item( int, const Py::Object & )
+int PythonExtensionBase::sequence_ass_item( Py_ssize_t, const Py::Object & )
 	{ missing_method( sequence_ass_item ); return -1; }
 
-int PythonExtensionBase::sequence_ass_slice( int, int, const Py::Object & )
+int PythonExtensionBase::sequence_ass_slice( Py_ssize_t, Py_ssize_t, const Py::Object & )
 	{ missing_method( sequence_ass_slice ); return -1; }
 
 
 // Mapping
-int PythonExtensionBase::mapping_length()
+Py_ssize_t PythonExtensionBase::mapping_length()
 	{ missing_method( mapping_length ); return -1; }
 
 Py::Object PythonExtensionBase::mapping_subscript( const Py::Object & )
@@ -1164,13 +1164,13 @@ Py::Object PythonExtensionBase::number_power( const Py::Object &, const Py::Obje
 
 
 // Buffer
-int PythonExtensionBase::buffer_getreadbuffer( int, void** )
+Py_ssize_t PythonExtensionBase::buffer_getreadbuffer( Py_ssize_t, void** )
 	{ missing_method( buffer_getreadbuffer ); return -1; }
 
-int PythonExtensionBase::buffer_getwritebuffer( int, void** )
+Py_ssize_t PythonExtensionBase::buffer_getwritebuffer( Py_ssize_t, void** )
 	{ missing_method( buffer_getwritebuffer ); return -1; }
 
-int PythonExtensionBase::buffer_getsegcount( int* )
+Py_ssize_t PythonExtensionBase::buffer_getsegcount( Py_ssize_t* )
 	{ missing_method( buffer_getsegcount ); return -1; }
 
 //--------------------------------------------------------------------------------
